% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{begin_redirect_stderr}
\alias{begin_redirect_stderr}
\title{Redirect stderr to a given file.}
\usage{
begin_redirect_stderr(filepath, append = FALSE)
}
\arguments{
\item{filepath}{The path to the desired file. Must not be null.}

\item{append}{Controls whether the text should be appended if the
file already exists. By default (\code{false}), truncates the file.}
}
\value{
An integer corresponding to the file descriptor of the old
stderr, so that the redirection can be later reverted.
}
\description{
This function redirects the standard error (stderr) to a new file
descriptor, pointing to the file at \code{filepath}.
}
\seealso{
\code{\link[=end_redirect_stderr]{end_redirect_stderr()}}.
}
