% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{end_redirect_stderr}
\alias{end_redirect_stderr}
\title{Reverts the redirection of stderr.}
\usage{
end_redirect_stderr(old_stderr)
}
\arguments{
\item{old_stderr}{The file descriptor to the old (previous) stderr,
before the redirection. This corresponds to the return value of
\code{\link[=begin_redirect_stderr]{begin_redirect_stderr()}}.}
}
\value{
The result code, usually equal to \code{2}. A negative value,
such as \code{-1}, indicates an error.
}
\description{
This function redirects the standard error (stderr) to its old file
descriptor (usually, this ultimately means the R Console).
}
\seealso{
\code{\link[=begin_redirect_stderr]{begin_redirect_stderr()}}.
}
